Move utils only for yolov7 and v9
