+++ src/otx/algo/detection/layers/elan_layer.py
@@ -13,7 +13,7 @@ from typing import Any, Callable
 import torch
 from torch import Tensor, nn
 
-from otx.algo.detection.utils.yolov7_v9_utils import auto_pad
+from otx.algo.detection.utils.utils import auto_pad
 from otx.algo.modules import Conv2dModule, build_activation_layer
 
 logger = logging.getLogger(__name__)
